pyproject.toml
[project]
name = "try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncio>=3.4.3",
    "openai>=1.79.0",
    "patchright>=1.52.3",
    "pydantic-ai>=0.2.4",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.1",
    "sqlmodel>=0.0.24",
]
